10.21~10.24/aws실습자료/aws/ec2/requirements.txt
flask==3.0.3
boto3==1.34.139
python-dotenv
pymysql